Allow providing additional types to schema build.
With the current API, only concrete instances which are registered
somewhere from the query or mutation root are registered.  This makes
writting tools to automatically generate types a PITA - specifically
because you have to ensure there is at least one concrete instance of
every type registered.

This change provides a way to pass a set of types into the schema when
building it, which will be added into the set of available types when
replacing type references.

An alternative could be to provide a resolver to GraphQLTypeReference,
which would be used to resolve the type at build time - however, this
method seems far less intrusive for existing consumers :-)