Allow keywords anywhere 'Name' is used.
The GraphQL specification (and reference implementation) doesn't
restrict the keywords 'query' or 'fragment' as Names.  Because of
ANTLRs behaviour, input with arguments, field names, or types with
either of these values were being rejected, e.g:

query Something {

  search(query: "theo") {
    ...
  }

}

This patch updates the grammar and the conversion to allow these
keywords.

Note that the gradle build scripts do not seem to be automatically
updating the generated source code, so i manually regenerated it.