Mess around with ANTLR grammar to get more Name.
The graphql-js parser is rolled by hand, so handles context of tokens a
lot better than ANTLR can.  Specifically, we need to ignore the 'on'
after '...' in the fragmentSpread, and can't use 'true' or 'false' even
through the graphql-js implementation allows it.

It might be easier to hand roll a parser for building the grammar rather
than use ANTLR for parsing as the GraphQL grammar isn't context free.

Perhaps somethign for another day.  For now, this works (except some
edge cases of argument names like 'true', false', etc).