ValuesResolver fixes for input object literals and variable values
Addresses #152 and #106

When resolving an object literal check for any fields which are omitted from
the literal but have a default value. In that case copy the default value to
the result. This fixes #152 (InputObjectType defaultValue is not getting used)

When resolving a variable value that corresponds to an input object type omit
any fields that are not present in the value, unless those fields have a
default value (in which case use the default value) or non null (which should
force an error). This fixes at least one part of #106, Unspecified fields in
Input object are set to null