Initial take on adding GraphQL Schema parsing.
The definitions come from this pending pull request:

https://github.com/facebook/graphql/pull/90

Specifically the grammar comes from:

https://github.com/facebook/graphql/blob/de5433206643b2ec9e87e18e142f2e630bfa05b9/spec/Appendix%20B%20--%20Grammar%20Summary.md

Note that the grammar is still not solidified, but tooling is already
being created that is based on this grammar. For example:

https://github.com/apollostack/graphql-tools

My hope is to enable schema-first development for Java by creating
tooling similar to the above, but for Java.

Note that I had to change the 'NAME' token to be a 'name' rule. This
was necessary to support "keywords" throughout the grammar. For
example, without this change there are tests that define "type" and
"scalar" as field names in various queries, and these would fail to
parse.