Make propertyName protected instead of private to allow extend-ability
Currently PropertyDataFetcher is not extendable since propertyName field is private. we have cases we'd like to extend it in order to use different getter algorithm on some fields but not all (e.g. one field is a jsonObject which we get values from).
Today we need to wrap the PropertyDataFetcher with our impelentation of DataFetcher and call PropertyDataFetcher in case source is not instance of JsonObject (similar to what you do with maps). making propertyName protected will make it cleaner - extend the class and call super.get()