Better performance for boolean properties with getXXX accessors (Kotlin)
Kotlin data classes generate getXXX methods for Boolean properties
instead of isXXX. This causes NoSuchMethodException to be thrown
on each access to a Boolean field, which is expensive performance-wise
due to the fillInStackTrace() method called when the exception is
created. Avoid this by caching the method prefix (get or is).