Add support for lazy lists
Lazy lists are lists whose elements are completed only when the user
tries to access them. The user can opt into this by returning an
instance of LazyList from their DataFetcher.

If the user uses lazy lists then they cannot use the toSpecification
method to get a specification-compatible representation of the execution
result. Instead they must manually serialize the result the result. The
reason for this is that the set of errors is only complete once the data
has been fully serialized.