Disallow complex input values to be coerced to String (#1665)
Number, String and Boolean are still allowed to be coerced to String
in input because those types have well-defined string representations.

However, passing values of complex types to String fields in input
values will result in CoercingParseValueException.

This behaviour is in line with current parsing logic for Boolean
and numeric types.

Note: serializing complex values (Map, Lists, custom classes) on
output is still allowed since the query implementation controls
the toString() representation in that case.