Add new schema printer option includeDirectiveDefinitions
Apollo Federation requires that we return the schema as part of the _Service field to help the gateway understand all the fields. See 'Fetch Service Capabilities' https://www.apollographql.com/docs/apollo-server/federation/federation-spec/#fetch-service-capabilities. This schema should not contain the diretive definitions but SHOULD contain all the directives on the types and fields. This PR adds a new option to the SchemaPrinter, includeDirectiveDefinitions, which is on by default. Setting it to false will remove the definitions from the top of the schema