#1839: Updated newly created IntrospectionUtils class
Description:
Updated the IntrospectionUtils class created in original PR to provide
different methods for testing introspection rather than having a single
monolithic method.  See below for details.

Changes:
* Introduced isIntrospectionOperationName that tests whether or not
an operation name is the traditional introspection operation name.  See
JavaDoc for more details.
* Introduced isIntrospectionQuery that tests whether or not
a query is an introspection query by checking for "__schema".  See
JavaDoc for more details.
* Introduced isIntrospectionDocument that tests whether or not
a GraphQL Document is an introspection query by checking for system
fields.  See JavaDoc for more details.
* Introduced isIntrospectionOperationDefinition that tests whether or
not a GraphQL OperationDefinition is an introspection query by checking
for system fields.  See JavaDoc for more details.
* Introduced isIntrospectionFieldSubSelection that tests whether or not
a GraphQL FieldSubSelection is an introspection query by checking for
system fields.  See JavaDoc for more details.
* Split the building of the IntrospectionQuery.INTROSPECTION_QUERY into
separate constants so that portions could be used to create constants
needed for unit testing.
* Created unit tests that check the validity of the new utility methods.
The unit tests also report the timings of the methods and have an option
for enabling testing the performance of said methods.

Unit Tests:
* Created unit tests that check the validity of the new utility methods.
* The unit tests also report the timings of the methods
* Unit tests have an option for enabling testing the performance of said
methods
* Performance testing is disabled by default.  The performance test
functionality is there as a sanity check that should be run whenever
the utilities are modified