Coerce the ConnectionCursor by using the interface
The DefaultConnectionCursor had a convenient toString()-method which used the value from
the getValue method in the interface ConnectionCursor. This made the ScalarsIDTest to
pass by accident. This commit will change the default toString()-method to be more aligned
with the other relay implementations and add a check for ConnectionCursor interface in the
ScalarID coercing code.