Add DelegatingDataFetcherExceptionHandler.Builder
This addresses two concerns:

The first is the use of LinkedHashMap for the delegates. LinkedHashMap was
chosen intentionally because the order matters. However, it is strange to
rely on an implementation. The Builder now encapsulates how the mapping
is done (using a List and private class named PredicateDelegateMapping).

The second concern it addresses is to ensure
DelegatingDataFetcherExceptionHandler is immutable allowing defaultHandler
to only be set in the Builder.