Fix interfaces implementing interfaces sub-type validation (#2434)
This commit fixes `TypeDefinitionRegistry.isPossibleType` and
`isSubType` to consider an _interface_ that implements another interface
as a sub-type of the implemented interface. Before this commit, only an
_object_ type implementing an interface was considered as a sub-type of
the implemented interface. So this small change brings the
implementation into full alignment with the below part of the spec:

IsValidImplementationFieldType(fieldType, implementedFieldType):
  ...
  5. If {fieldType} is an Object or Interface type and
     {implementedFieldType} is an Interface type and {fieldType}
     declares it implements {implementedFieldType} then return {true}.
  ...