Ensure field errors are populated in FetchedResult before instrumentation `beginFieldComplete` is called
Currently if a DataFetcher raises an exception in any way, the FetchedValue given in the
instrumentation does not contain any errors information. The only place errors are available
are on the ExecutionContext within a list, which is not ideal to determine if the current
field has an error. This change ensures that FetchedValue contains errors if any was raised
during `DataFetcher.get`.