Gracefully returning null in cases of UnresolvedTypeException
The different code path that existed prior to this change would omit the calls to `onFieldValuesInfo` when an `UnresolvedTypeException` occurred, and that causes a thread leak similar to an older issue for unhandled exceptions https://github.com/graphql-java/graphql-java/issues/2068 - that is, once a type cannot be resolved, the deeper batches from other (successful) branches are never dispatched.

Since we are just adding an error in the cases of `UnresolvedTypeException` and try to set the value to null, this only needed a small refactoring to make it similar to when the value is actually null.