Use stable version number on clean workspace
Previously, the graphql-java build would always produce a new version
number based on the current date and time in local development. Because
the version number would always be different between 2 Gradle
invocations, some tasks, such as `javadoc` or `jar` could never be
cached because they depend on the version number.

With this patch, the version number will no longer include the date and
time when the Git workspace is clean, because the commit hash is
sufficient to uniquely identify the state of the workspace. This allows
the tasks that depend on the version number to be properly cached.