Relativize Antlr sources in cache key
Previously, the cache key of the antlr task would be computed using
(among other things) the absolute paths of its input files. Because the
absolute paths were part of the cache key, the build cache would rarely
be hit.

With this patch, the task is configured to use relative paths
when computing the cache key.