Implement equals/hashCode for GraphqlErrorImpl
We have a great many tests that verify that the errors emitted from a
`DataFetcher` fit a certain shape.

However, verifying equality of these errors is fiddly and error-prone,
as we must directly check each individual attribute on every error -
this is painful especially when we are trying to perform assertions on a
`List` of `GraphQLError`s.

To this end, we add `#equals` / `#hashCode` methods on
`GraphqlErrorImpl`. However, it is important to note that `equals` will
return `true` if all the attributes match, even if the implementing
class is _not_ a `GraphqlErrorImpl`. This is done so that different
implementations may be swapped in and out with causing test failures.