Add test reporting to pull request workflow
On a build with failing tests it is currently hard to see exactly which
test failed.

We thus add a popular test reporter to publish JUnit test results so
that they can easily be examined.

As workflows need to be approved by the maintainer to run, we will only
be able to confirm whether this works once the maintainer enables this
themselves.